Made preprocessor more robust
The preprocessor will now eliminate more false matches of the log
function. However, it will still fail if the user invokes log
through a pointer or they mask the log function with a custom one.
In the latter case, the preprocessor will attempt to rewrite that
function and crash.

This commit is also done in preparation of re-writing the parser
such that it starts embedding logging code rather than just
perform a text replacement of the LOG message + fmtId.