WIP - First working prototype with arbitrary args
FastLogger now supports printf-like semantics with arbitrary format strings
and arguments. The only format argument to not be supported is "%n".

That being said, this version is still very cludgy and needs to be cleaned
up. There exists untested code, broken test code, missing comments, and
strange abstractions that need to be reworked.