Fixed a rollover bug in StagingBuffer
As noted in the code comments, we define the buffer to be completely
empty when the consumer and producer positions are equal to each other.
Unfortunately, there was a bug whereby the producer would roll-over the
buffer without checking the consumer position, thus overlapping the two
values if the consumer was also at the beginning of the buffer. The fix
was to check the consumer position before rolling over as the producer.