Fixed a compilation errors in generated code
FunctionGenerator.py generated syntactically incorrect code in the form
of "size_t size_t str1len;" when it encountered multiple "%s" parameters
in the user NANO_LOG() invocation. This is now fixed and tested for in
the integration tests.

Also changed the type for "%p" to "const void*" instead of "void*" for
const correctness.