Refactored the NanoLog interface code
NanoLog and NanoLog's internals now exist in different namespaces,
so they should no longer conflict with user code. In addition,
NanoLog.h is now cleaner; it only contains the user interface and
is well documented.

Past users of NanoLog will need to update their GNUMakefile to
include a new internal class RuntimeLogger.cc in the RUNTIME_CC
macro and also include the LogLevel namespace (it is no longer
global) in their sources. See the changes in the benchmark folder.