Added C++17 (Non-Preprocessor) NanoLog
The repository now contains an alternate implementation of
NanoLog that uses C++17 features rather than a preprocessor
to analyze and slim down log statements at compile-time.

This makes NanoLog much easier to integrate into other
projects (just include the header and link the library).
A sample project demonstrating how to use this version of
NanoLog is provided in /sample.