Modified GNUmakefiles to fix unit tests
The last two commits broke the compilation of some unit tests by
adding #ifdef's that compiled out some logic in the NanoLog library.
This commit fixes the issue by modifying the GNUmakefiles so that
the right version of the library is compiled for C++17 NanoLog,
Preprocessor NanoLog, and the unit tests.